requirements.txt
github-copilot-sdk
PyYAML
